+++ CHANGELOG.md
@@ -1,4 +1,7 @@
+## 1.2
+Add reload method(), fix meta data login
+improve code struct
 ##1.1
-Edit code struct, improve logic
+Edit code struct, improve code logic
 ##1.0
 Create Project